cloudshell-L1-{{cookiecutter.project_slug}}/requirements.txt
cloudshell-core==2.2.176
-e git+https://github.com/QualiSystems/cloudshell-L1-networking-core.git@refactoring#egg=cloudshell_l1_networking_core
